pyproject.toml
[project]
name = "general-analytics"
version = "0.1.0"
description = "Multi-client GTM & Google Analytics Management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-analytics-admin>=0.27.0",
    "google-analytics-data>=0.20.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "supabase>=2.27.2",
]

[tool.uv]
dev-dependencies = []
